From http://www.jwz.org/xscreensaver/xscreensaver-5.27.tar.gz
-rw-r--r-- 1 zblaxell zblaxell 8786914 May 27 06:18 xscreensaver-5.27.tar.gz
20c4c0f501ef92b2752397c0f25ce4c0bd5bcde0  xscreensaver-5.27.tar.gz

+++ OSX/iosgrabimage.m
@@ -56,7 +56,7 @@ ios_random_image_done (ios_loader_data *d, BOOL ok)
 {
   UIImage *img = 0;
   const char *fn = 0;
-  int n = ok ? [d->assets count] : 0;
+  NSUInteger n = ok ? [d->assets count] : 0;
   if (n > 0) {
     ALAsset *asset = [d->assets objectAtIndex: random() % n];
     ALAssetRepresentation *rep = [asset defaultRepresentation];
